// src/interfaces/chat.ts
export interface ChatMessage {
  createdAt: Date;
  groupInfo: ChatInfo;
  image: null;
  isAutoGenerated: boolean;
  isGroup: boolean;
  message: string;
  receivers: string[];
  senderId: string;
  senderInfo: ChatInfo;
  views: string[];
}

export interface ChatInfo {
  avatar: string;
  desc: string;
  id: string;
  name: string;
}

export interface UserChatMessage {
  latestMessage: ChatMessage;
  unread: number;
}